requirements.txt
streamlit>=1.28.0
numpy>=1.24.0
pandas>=2.0.0
pandas-ta-classic
plotly>=5.18.0
python-dotenv>=1.0.0
yfinance>=0.2.28
line-bot-sdk>=3.0.0
FinMind>=1.0.0
pytest>=7.0.0
flask>=3.0.0
requests>=2.31.0
